requirements.txt
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pkg_resources==0.0.0
python-dotenv==0.19.2
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.20.2
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uvicorn==0.17.6
